requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.21.0
plotly>=5.0.0
seaborn>=0.11.0
matplotlib>=3.5.0
openpyxl>=3.0.0
xlrd>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
altair>=4.0.0
streamlit-option-menu>=0.3.0
streamlit-aggrid>=0.3.0
scikit-learn>=1.0.0
scipy>=1.7.0